hx-20

+++ src/machine/hx20.h
@@ -60,7 +60,7 @@ private:
     int int_status;
     int int_mask;
     int key_intmask;
-    int key_data;
+    UINT16 key_data;
 
     Cmc6800 *pmc6301,*pSlaveCPU;
